requirements.txt
Flask
flask-cors
torch
torchvision
Pillow
numpy
opencv-python-headless
